+++ cuda_lin_alg/src/cuda_utils.h
@@ -1,6 +1,6 @@
 #pragma once
 
 extern "C" {
-double* transfer_to_cuda(const double* host_array, const size_t count);
-double* transfer_from_cuda(const double* device_array, const size_t count);
+float* transfer_to_cuda(const float* host_array, const size_t count);
+float* transfer_from_cuda(const float* device_array, const size_t count);
 }
